Reduce dependencies on the library by moving validity and loading to the running module. - Now, only one dependency on the library remains. - It will fall soon.
