+++ server/modules/module_library.js
@@ -15,13 +15,7 @@ limitations under the License.
 
 import EventEmitter from 'events';
 import {EmptyModuleDef} from './module_def.js';
-import assert from '../../lib/assert.js';
 import {easyLog} from '../../lib/log.js';
-import path from 'path';
-import {Server} from '../../lib/module_interface.js';
-import conform from '../../lib/conform.js';
-import inject from '../../lib/inject.js';
-import * as wallGeometry from '../util/wall_geometry.js';
 
 const log = easyLog('wall:module_library');
 
@@ -34,50 +28,9 @@ class ModuleLibrary extends EventEmitter {
   register(def) {
     log.info('Registered', def.name);
     this.modules[def.name] = def;
-    if (def.serverPath) {
-      // Validate the module at the server path.
-      this.loaded.set(def.name, this.extractServerClass(def.name, {
-        network: {},
-        game: {},
-        state: {},
-      }).then(() => {
-        log.debugAt(1, 'Verified ' + path.join(def.root, def.serverPath));
-        this.valid.set(def.name, true);
-      }, err => {
-        log.error(err);
-      }));
-    } else {
-      this.valid.set(def.name, true);
-      this.loaded.set(def.name, Promise.resolve());
-    }
   }
   reset() {
     this.modules = {'_empty': new EmptyModuleDef};
-    this.loaded = new Map;
-    this.valid = new Map;
-  }
-  whenLoaded(name) {
-    return this.loaded.get(name) || Promise.reject(new Error(`Unknown module ${name}`));
-  }
-  isValid(name) {
-    return this.valid.get(name) || false;
-  }
-  async extractServerClass(name, deps) {
-    const def = this.modules[name];
-    const fullPath = path.join(process.cwd(), def.root, def.serverPath);
-    const {load} = await import(fullPath);
-
-    // Inject our deps into node's require environment.
-    const fakeEnv = {
-      ...deps,
-      wallGeometry: wallGeometry.getGeo(),
-      debug: easyLog('wall:module:' + name),
-      assert,
-    };
-
-    const {server} = inject(load, fakeEnv);
-    conform(server, Server);
-    return {server};
   }
 }
